# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Internal service for monitoring system status and tracking incidents",
        "title": "Status Incident API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/systems": {
            "get": {
                "description": "Get a list of all systems",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "systems"
                ],
                "summary": "List all systems",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.System"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new system with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "systems"
                ],
                "summary": "Create a new system",
                "parameters": [
                    {
                        "description": "System data",
                        "name": "system",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.createSystemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.System"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            }
        },
        "/systems/{id}": {
            "get": {
                "description": "Get detailed information about a specific system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "systems"
                ],
                "summary": "Get a system by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "System ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.System"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            }
        },
        "/systems/{id}/status": {
            "post": {
                "description": "Update the status of a system (green, yellow, red)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "systems"
                ],
                "summary": "Update system status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "System ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.System"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Status": {
            "type": "string",
            "enum": [
                "green",
                "yellow",
                "red"
            ],
            "x-enum-varnames": [
                "StatusGreen",
                "StatusYellow",
                "StatusRed"
            ]
        },
        "domain.System": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "description": "responsible person/team",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.Status"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "description": "link to the system",
                    "type": "string"
                }
            }
        },
        "http.createSystemRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "http.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "http.updateStatusRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}